Change data types of PurchaseCost and UsageCost properties to decimal

+++ DAL/Models/Statistics.cs
@@ -14,7 +14,7 @@ public class Statistic : Document
     [BsonElement("CreatedDate")]
     [BsonDateTimeOptions(Kind = DateTimeKind.Local)]
     public DateTime CreatedDate { get; set;}
-    public double PurchaseCost {get; set;}
-    public double UsageCost {get; set;}
+    public decimal PurchaseCost {get; set;}
+    public decimal UsageCost {get; set;}
     public int? NumberOfDays {get; set;}
 }
